refactor: organize existing classes in a new structure
