+++ app/src/main/java/com/picpay/desafio/android/utils/UserListDiffCallback.kt
@@ -1,7 +1,7 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.utils
 
 import androidx.recyclerview.widget.DiffUtil
-import com.picpay.desafio.android.User
+import com.picpay.desafio.android.data.api.responses.User
 
 class UserListDiffCallback(
     private val oldList: List<User>,
